todoapp/requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.7
django-dotenv==1.4.2
django-heroku==0.3.1
djangorestframework==3.12.4
gunicorn==20.1.0
Markdown==3.3.4
psycopg2==2.9.3
pytz==2021.1
sqlparse==0.4.2
whitenoise==5.3.0
